fix: copied wikilinks being encoded & undefined headings (#177)
WikiLinks do not need to be URI encoded. This change will ensure that WikiLinks are thus created in the unencoded form whereas Markdown Links (which include the `obsidian://` URL scheme) keep the URI encoding for the file path while also using the unencoded version for the Display. 

Also, I always got the `#heading` version, even if no heading was selected in Alfred, with a string value of `undefined`. This contains an attempt to fix this by testing for "heading" and its value.